+++ src/resqml2_0_1/DiscreteProperty.cpp
@@ -171,11 +171,11 @@ bool DiscreteProperty::validatePropertyKindAssociation(gsoap_resqml2_0_1::resqml
 	return true;
 }
 
-int64_t DiscreteProperty::getNullValue(unsigned int patchIndex) const
+int64_t DiscreteProperty::getNullValue(uint64_t patchIndex) const
 {
 	gsoap_resqml2_0_1::_resqml20__DiscreteProperty* prop = static_cast<gsoap_resqml2_0_1::_resqml20__DiscreteProperty*>(gsoapProxy2_0_1);
 
-	auto abstractIntArray = prop->PatchOfValues[patchIndex == (numeric_limits<unsigned int>::max)() ? prop->PatchOfValues.size() - 1 : patchIndex]->Values;
+	auto abstractIntArray = prop->PatchOfValues[patchIndex == (numeric_limits<uint64_t>::max)() ? prop->PatchOfValues.size() - 1 : patchIndex]->Values;
 	if (abstractIntArray->soap_type() != SOAP_TYPE_gsoap_resqml2_0_1_resqml20__IntegerHdf5Array) {
 		return (std::numeric_limits<int64_t>::max)();
 	}
